Moved default-cards app to lib/mobiledoc
refs #9178, refs https://github.com/TryGhost/Ghost/pull/9338
