+++ core/server/lib/mobiledoc/cards/image.js
