+++ core/server/lib/mobiledoc/cards/markdown.js
