+++ core/server/lib/mobiledoc/index.js
