+++ core/test/unit/lib/mobiledoc/atoms/soft-return_spec.js
@@ -1,5 +1,5 @@
 var should = require('should'), // jshint ignore:line
-    card = require('../../../../../server/apps/default-cards/atoms/soft-return'),
+    card = require('../../../../../server/lib/mobiledoc/atoms/soft-return'),
     SimpleDom = require('simple-dom'),
     opts;
