+++ core/test/unit/lib/mobiledoc/cards/hr_spec.js
@@ -1,5 +1,5 @@
 var should = require('should'), // jshint ignore:line
-    card = require('../../../../../server/apps/default-cards/cards/hr'),
+    card = require('../../../../../server/lib/mobiledoc/cards/hr'),
     SimpleDom = require('simple-dom'),
     opts;
